noria: Remove LocalOrNot
As far as we're aware, LocalOrNot<T> only ever existed to allow
artificial benchmarks for the Noria thesis to make faster writes by
sending raw pointers within the same process, rather than having to go
over a channel. In fact, the only entrypoint to create a "local" version
of a LocalOrNot was to call `Table.i_promise_dst_is_same_process()`,
which only ever happened in code in noria/applications, which is now
gone.

To avoid having to keep around potentially highly unsafe code (a
carefully crafted request to a Table RPC server could hypothetically
force it to access arbitrary memory locations, which is... not great!)
this commit gets rid of LocalOrNot entirely, instead just replacing it
with whatever inner type it was wrapping everywhere it was used
previously.

Death to unsafe code!

Fixes: ENG-113
Change-Id: I66c4b77c0b7092d95aa6d903348e80d786daec77
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2780
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Tested-by: Buildkite CI