Name nodes with Table throughout migrations and dataflow
In preparation for supporting multiple schemas throughout ReadySet, this
commit updates basically everywhere we name nodes to use a Table (which
consists of a name, and an optional schema) instead of just a
string-like name. In MIR, this replaces the previously-introduced
`Relation` struct, since it's now essentially identical - that struct is
now gone.

Probably most notable in this change is that this /does/ include the
name of dataflow nodes, in addition to just MIR nodes - we went back and
forth a fair bit on that decision, and ended up with this approach for a
few reasons:

- Perhaps unfortunately, lots of things use the names of dataflow nodes
  to index into other structures which *do* need schemas - for example
  when we look up schemas for tables which we've obtained via only a
  node index - and it's simpler to avoid having to add extra
  "table_name: Table" fields or something like that to avoid that extra
  information
- Naturally, it's simpler to just keep a full Table through the
  mir->flow conversion rather than trying to make up a string by
  stringifying the schema and table name
- Conceptually, it kind of makes sense that if relations in a database
  "live inside" schemas, the nodes implementing those relations would as
  well.

Finally, since we now need the correct schema involved in table creation
and lookup, this does the rest of the necessary threading of the schema
search path through migrations (to the new rewrite pass which resolves
schemas on tables in statements) and through replicators (both when
creating tables and when looking them up for their mutator)

Since these changes slightly change the hashes generated for
queries (and hence the names for nodes) lots of the assertions in
controller/sql/mod.rs broke - I'm not personally crazy about those
tests, in part because of how brittle they are in cases like these, and
I'm not sure they're covering anything valuable that isn't otherwise
covered by a different test, so in this case I've just opted to remove
the failing assertions.

Another notable change is that lots of functions to make mir nodes that
previously took their `name` argument as a reference now take that
argument owned - this is because basically all of those functions just
cloned that argument anyway, so it's better style to pass it owned.

Fixes: ENG-1580
Change-Id: Ia9ad20573d2fceebc1a155cb2995e34a22858a29
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2787
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>