nom-sql: Make cache names Tables
Because of the way we track cached queries and views, it's
actually *easier* to allow the names supplied to `CREATE CACHE` and
`DROP CACHE` statements to be qualified with a schema than it would be
to go through the extra work to omit the schema. This commit starts to
lay the groundwork for allowing that, by making the `name` field on the
CreateCacheStatement and `DropCacheStatement` AST structs use `Table`
instead of  `SqlIdentifier`. That's basically all that's happening here
- everywhere else we'd need to care we're just ignoring that schema for
now, and a TODO comment has been left.

Change-Id: I8c57825322745992834d7b390d4952ed17b8ce85
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2791
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>