mir: Create new petgraph::Graph MIR representation.
As a first step for what will be a huge refactor, this commit
introduces the new structures that will be the new foundation for MIR.
The new `MirGraph` is meant to replace the old, distributed MIR
graph that is represented by a mesh of `MirNodeRef`s.
The methods implemented on this struct are currently not used, and
will be used in future commits.
This also adds the new `MirGraph` to the `SqlToMirConverter` (unused for
 now).

The `MirGraph` exposes generic graph methods (by allowing users to
deref it and work directly with the petgraph's `StableGraph` underlying
structure), it also provides functionality specific to MIR, such as the
`columns` method that allow us to resolve the columns outputted by a
node.

Refs: ENG-1622
Change-Id: I22157eeef752ee01f66066bc41df9e575e520181
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2957
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>