nom-sql: add position information to parsing errors
This commit leverages the nom_locate crate to provide line and column
numbers to SQL parsing errors. It does this by wrapping all input
sequences in nom_locate::LocatedSpan, and implementing a custom error
type for nom::IResult that propagates the error returned by the parsing
operation that made it the furthest, e.g. out of the branches of an alt
combinator.

Co-authored-by: Dan Wilbanks <dan@readyset.io>
Change-Id: I12a6ae2f3c62d28f2236ed0ed2f3a32eb472ba54