nom-sql: Don't do desugaring in parser for CREATE TABLE
Doing the kind of normalizing AST rewriting we were doing in the CREATE
TABLE parser was always kind of wrong, primarily because it's generally
not a great idea for that kind of rewriting to happen in parsers in the
first place - but as we're preparing to add support for multiple schemas
throughout the system it added a reasonable amount of extra complexity,
so now feels like a reasonable time to get rid of it.

This commit does that, rewriting the parser to just be a normal
parser (using sequenced `let` instead of `tuple` abuse while we're at
it and renaming the parser from "creation" to "create_table", to make
things more readable and maintainable). Now, the bit that we *really*
care about, which is adding the table to the Columns mentioned in the
statement, is done via a (much simpler!) rewrite pass using a Visitor.

Change-Id: I0a83a53dd5af149a1cb067d8557dc296f3802953
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2807
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>