Move expression lowering into dataflow_expression
There are things other than the MIR->Dataflow lowering step which will
want to be able to turn `nom_sql::Expr`s into
`dataflow_expression::Expr`s - specifically, in this case, I'm preparing
to write a constant folding pass, which will do this lowering pass so it
can actually *evaluate* constant expressions. To allow for that, this
commit moves that expression lowering step into the
`dataflow_expression` crate, as an associated function on Expr which
takes the nom_sql Expr, and a function to resolve a named column
reference into a numeric column index. Since it's now broken out as a
nicely isolated function, this also takes the opportunity to cover it
with a quick test.

Change-Id: I510087d6ba64e75a9f344ce04311691c61ebf489
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2813
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>