sql-passes: Add a constant folding pass
Especially with things like ORMs that abstract over programmatic
construction of SQL queries, it's common to end up with expressions or
subexpressions in queries that're constant-valued - for example,
to avoid the overhead of stripping off an extra `AND`, wordpress creates
queries with conditions that look like `WHERE 1=1 AND author_id = ?`.

As an optimization, this commit adds a new constant folding rewrite pass
on expressions, which would turn that into `WHERE 1 AND author_id =
?` (with `1` standing in for "true"). Later, I'd like to also add a dead
code elimination pass, which would know about the identity elements  for
various binary operators and could turn that into `WHERE author_id = ?`.

Under the hood, this works by translating any constant-valued
subexpressions anywhere in the query into dataflow expressions,
then *evaluating them* using the same expression evaluator that we use
inside of dataflow, and converting the resulting `DfValue` back to a
`Literal` to put into the query. This may feel like it breaks the layer
boundary between components a bit, but has the advantage of guaranteeing
no bugs caused by a difference in semantics between a hypothetical
separate constant folding evaluator and the dataflow expression
evaluator - plus there's prior art! See Normalization By Evaluation[0]
for more.

To keep rewrite passes that only normalize expressions both colocated
and simpler to write, this introduces a new `expr` module to
readyset-sql-passes containing this pass, and moves the previous
negation normalization pass in there.

[0]: https://en.wikipedia.org/wiki/Normalisation_by_evaluation

Change-Id: I91cbc63ccd3eed26f5596261a0305c7b364ab678
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2814
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>