MySQL: Return Display Length for Column when available
mysql-srv previously hardcoded the display length for all columns to
1024. This causes a problem for the Ruby mysql2 driver, which relies on
the display length field of TinyInt to determine whether the column
represents a boolean.

This CL adds an Option<u32> column_length field to mysql-srv::Column
which preserves the display length of the column from the SqlType or
mysql-async::Column. If the display length was not available in SqlType
we will default to 1024. The display length is always present in
mysql_async::Column, so we return the correct value for proxied queries.

The column_length field should be changed from Option<u32> to u32,
because we must encode a display length for every column we return.
However, the CL in its current form provides a simple fix to return the
correct display_length when the value is explicitly defined.

We likely want to replace mysql-srv::Column with mysql-async::Column,
because the latter contains all the fields needed to encode a column in
the resultset.

This CL also adds a link to the MySQL com-query-response docs.

Verified the fix with ruby using the following script

````
require 'mysql2'

con = Mysql2::Client.new(:host => "127.0.0.1", :username => "root",
:database => 'noria', :port => '3307', :password => 'noria')

result = con.query('SELECT * FROM t;', :cast_booleans => true)
p result.first
````

Change-Id: Ide42bf8f481642144a57a1f348d0347b8ca67c4c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2822
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>