mir: Refactored MIR to use petgraph instead of MirNodeRef.
This commit changes the current MIR graph implementation:
+---------------+
|    Before     |
+---------------+
The MIR graph was represented through `MirNodeRef`s, which were
references to `MirNode`s, and where each node had the references to its
ancestors and descendants.
On top of that, if we wanted to reuse a new node, we would create a new
one that would have a separate reference to the original node (which
would also make possible to reuse a reused node!), and which type was
considered to be `MirNodeInner::Reuse`.

That representation worked also to create different graphs for each
`MirQuery`. Those queries were disconnected from each other (each of
them was _supposed_ to be self-contained), meaning that their nodes
would only have references to other nodes of the same query.
The only connection between them was through the separate reuse
reference.

This representation also required custom serdes algorithms to
reconstruct the references, since the circular references between
ancestors and descendants would not allow us to simply derive
`Serialize`/`Deserialize`, because the serdes would loop forever.

Another thing to note, is that the way a node held the references to its
 ancestors was important: that same order was assumed to be the order
 the ancestors should be resolved.
This is specially important for JOINs and UNIONs, for column resolution.

+---------------+
|     After     |
+---------------+
The MIR graph is represented using petgraph's `StableGraph`.
The graph itself holds the relationship between the nodes, simple as
that.
In order to support reuse, the nodes now hold a new field called
`owners`, which we use to effectively keep track of which queries own
it.
This new concept of node ownership allow us to:
1. Easily reuse a node, just add the query to the set of owners!
2. Correctly handling the scenarios where we remove a query: we just
remove the query from the set of owners, and if the set of owners is
empty (and if the node is not a base node, since those should NEVER get
deleted), then it's safe to delete the node!

To solve the problem of column resolution (ordered ancestors), we store
a `usize` as weight for the edges between nodes. That `usize` will
essentially serve as the ancestor index.

Change-Id: I33b818997f97f1b6c3f6ec14d05026de32a27bf5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3346
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>