show proxied includes parse failures
Prior to this change, `SHOW PROXIED QUERIES` would not show queries that
are proxied due to a parsing failure.

This changes Query to be an enum rather than just a type alias for
SelectStatement. Most changes are encapsulated within QueryStatusCache,
allowing callers to still pass in SelectStatements to functions there.

The Arc<Query>s in QueryStatusCache's DashMaps were moved into the enum
itself. This makes the enum variants equally sized and is required to
cleanly satisfy the DashMap Borrow<Q> trait bound on its getter fns.

Refs: ENG-1079
Change-Id: I6479245fed33434fee4a2db3e5bff119aedb6a77
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2816
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>