dataflow: remove max_concurrent_replays
It is not entirely clear what role this setting serves, but it
was introduced more than five years ago, and doesn't seem to do
anything useful nowadays. It was probably supposed to throttle
the number of messages passed around, but the default is set so
high it is unlikely to ever be saturated to begin with. Other
than that the messages are processed serially anyway so a natural
throttling is always in place. If it is meant to prevent too many
RequestPartialReplay messages blocking a domain, then we have a
much more efficient way to reprioritize messages in the 'receive_packets`
method.

Change-Id: Ie21d1a0aac71c0c5982fe8d41fd7b48b9c916d76
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2835
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI