treewide: Update rocksdb to v0.19.0
Update rocksdb to v0.19.0 to fix RUSTSEC-2022-0046. We're not affected
by that, since it only affects a method we're not using, but we might as
well update.

Only one actual code change was required here: the item type for
`full_iterator_cf` now returns a Result, but since we've generally been
ok with just unwrapping stuff in persistent state (since all the errors
are basically unrecoverable for us right now) this just unwraps that
result in a `map`.

Also, readyset-dataflow was depending on rocksdb but not actually using
it; this just removes that dependency.

Change-Id: I2dd81491502f327e1c9a77fcab0411cce8cf5656
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2836
Tested-by: Buildkite CI
Reviewed-by: Ian Meyer <ian@readyset.io>