dataflow: avoid cloning and allocations on retain
Retain + KeyComparison::contains are used quite a lot
in the replay path, however for some reason it was
cloning and collecting the keys multiple times, even
though it wasn't really needed. This change fixes this
problem. In a range-query replay-heavy flamegraph the
cost of retain is now reduced from 40% to just 1.3%.

Change-Id: Icc2deed4f17988871179a542995a7f6e1b60d0d4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2845
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI