Extend coerce_to function to take a "from" type
The frequently-used coerce_to function only takes a SqlType for the type
that you want to convert the value *to*, which is predicated on the
assumption that DataType contains all the type info needed to do this
conversion. While this has been more or less true in the past, it is
entirely possible to have different conversion semantics for different
source types that may share the same underlying DataType representation.
To allow for that, this commit adds a parameter to coerce_to to indicate
which SqlType we are converting *from*.

Note that from_ty is DfType instead of SqlType, because in many cases we
don't have an underlying SqlType (such as when we coerce a string
literal to whatever other type, rather than a value from a column with
an established type). In this case we pass in DfType::Unknown, and
should be able to fall back on the existing coercion semantics.

This change should make no functional difference for now, since it does
not actually add any changes in the type conversion logic, but it will
be useful in the near future when we add support for enum types.

Change-Id: Ia5b26eb3cdfcb9ebe7b97a8ebb247786f947e947
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2849
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>