Nom-sql: Parse and drop sign for Float and Decimal
Updates nom-sql to check for a SIGNED/UNSIGNED keyword after float and
decimal. The keyword is parsed, but dropped. The UNSIGNED keyword does
not change the behavior of floats or decimals, but indicates that
negative values should not be inserted into those columns. Users
connected to their upstream database will have this check enforced by
their database, so we can safely ignore these keywords.

Change-Id: Ie869145c475ea9056de88b28404ad7b71089127a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2855
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>