readyset-data: Add preliminary support for enum types
This is still a work in progress, and has enough missing pieces that it
is not really useful on its own, but it does compile and work and will
act as a stepping stone for later commits.

Although this doesn't touch the adapters yet, the type we are adding is
meant to implement the MySQL enum semantics. When we eventually add
Postgres enum support, we will either write shim code in the adapter to
translate the representation and use the same code under the hood as
MySQL, or we will add a separate Postgres-specific enum type to
readyset-data.

As an alternative, we might have been able to build enums off of
existing integer variants of DfValue, and done the conversion to enum
representations further out at the edges of the graph when we are
returning results, but adding a new DfValue variant seemed like a more
straightforward approach.

Change-Id: Ie347503a44eeca626903c0b5b860c905addfee01
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2866
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>