readyset-data: Add ability to convert text types to enums
This is necessary to correctly support MySQL enum semantics. Note that
this might not be super efficient for very large enum types (i.e. enums
with many possible values) since we're doing a linear search through the
Vec, but I don't think it should be terribly hard to optimize this in
the future by just changing SqlType to use a different data structure
for Enums.

As with the prior commit implementing enum/integer coercions, this
change is purely intended to implement MySQL enum semantics. Postgres
support is still being planned as a future enhancement.

Change-Id: Idfaf5096061b323f92ed253fd0f5d99c93282522
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2868
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>