psql-srv: Support Void PassThrough results in Text protocol
PassThrough is not yet implemented for postgres text resultsets. This
adds a special case for PassThrough values with a type of Void. This
type indicates that a function call returns no value, which we can
handle.

This case was hit in practice when attempting to process a 'SELECT
pg_advisory_lock()' command.

Change-Id: Iaecbe06cca3c91bf6c30d363bf07c19c25bdb848
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2885
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>