Fix UB from `DfValue::eq` fast path
Enum variants like `None` and `Max` have uninitialized bytes, which are
not allowed to be directly read. While the compiler is likely to give us
the expected correct behavior here, it is not required to. It may even
return `true` for two different enum variants since it may assume this
read operation will never happen.

Change-Id: I59d984d754b4b3f4a579f369b3b5c7a24220e13d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2889
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>