Deduplicate identical keys in multi-key lookup
Ignores duplicate keys in a multi-key lookup. The predicate of a SQL
query specifies the set of rows to select. Rows that satisfy multiple
conditions in the predicate should only be returned once.

Previously, if we received a query that contained the clause 'WHERE foo
IN (1,1)', we would return a resultset that contains two copies of all
the rows satisfying 'WHERE foo = 1'. We should only return a single copy
of each row in this case. This is fixed by deduplicating keys in the
reader.

We already deduplicate keys for range queries, so this only affects
point lookups.

Change-Id: I37dfe8444dfb7c71674d336639e310e431a7f813
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2898
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI