replicators: ignore failed rows for filtered tables
Currently when WAL fails to parse a row for an unsupported type
the replicator process panics. This change instead returns an
error from the replicator, and in addition it checkes the table
for which an error occured, and if it is for a table that would
be ignored anyway the error is skipped over.

Change-Id: Ifcd508203d71bfc26ace8fdb2b294fb5884d491f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2902
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>