client-adapter: Tweak consul init poll frequency
In the readyset-client-adapter, poll more frequently to try to establish
a consul session if we don't have an active one. If we do, use the
existing 20 second poll interval.

This reduces the time it takes for a ReadySet k8s deployment to become
healthy and I imagine would help reduce unavailability if the consul
session needs to be re-established.

This change also adds more logging around adapter initialization that
was useful in determining where it was getting stuck. Since these logs
occur once per process, it makes sense to be more verbose to help debug
startup issues.

Change-Id: Iad49642fcb0c94771d3b0a10375297b1b2001157
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2899
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>