nom-sql: Parse postgres escape character
In postgres, an "E" preceding a single-quoted string indicates that
escape sequences should be recognized in the string. This CL adds
support for parsing and ignoring this preceding escape indicator.

Currently, we recognize escape sequences in all postgres strings, so
ignoring the preceding "E" is fine. However, if we ever allow users to
toggle the standard_conforming_strings setting, the presence of an "E"
should change our behavior.

Change-Id: Ibf8477799d4bb9ff0be9804b312adbc5717ac8dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2912
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>