ALWAYS should mean always
This commit changes the behavior of the optional ALWAYS flag to actually
mean always. It now ensures that a query cached with the ALWAYS flag
will also bypass explicit transactions. It also changes behavior so we
don't require successful migration - we should instead return an error
if ReadySet fails to handle the query.

Change-Id: Ifada52156778bbb79d756b9cad6352f427bd0fb4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2911
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>