nom-sql: Parse ONLY in ALTER statement
Postgres allows the ONLY keyword in ALTER statements to indicate whether
the alter statement should affect tables that inherit from the table
mentioned in the statement.

We do not support table inheritance, so we may ignore this keyword.
However, the presence/absence of this keyword is stored in the
AlterTable struct for future use.

MySQL does not use the ONLY keyword in ALTER statements, though the
keyword is reserved.

Change-Id: I561703e2eb25c21b0ccdee8556d3dfbdf8b750bb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2916
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>