nom-sql: Support ALTER TABLE <table> DROP CONSTRAINT
Adds support for DROP CONSTRAINT in ALTER TABLE. Does not add support
for ADD CONSTRAINT, which is slightly more involved.

This leaves the implementation of handling DROP CONSTRAINT in the recipe
unsupported. However, this adds DROP CONSTRAINT statements to the list
of DDL that require resnapshot, which allows us to support these
statements when connected to an upstream database.

Change-Id: Ice54ef53f12d4a611c17486f5d7ee7d554650404
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2917
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>