readyset-data: Implement missing Ord clause for DfValue::Enum
The lack of this implementation was easy to miss because right now we
serialize enum values as integers, so after they are retrieved back from
RocksDB we are no longer using the Enum variant of DfValue. That is
about to change in an upcoming commit, though, so this commit is needed
to ensure proper ordering of enum values in query results.

Change-Id: Id87e0d381916e57f84bf0c0c7832b5cb07df508e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2928
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>