refactor: Eliminate the DfValue::Enum variant
After some discussions with Griffin, we've decided it's not necessary to
have a separate DfValue variant for enums. All of the same functionality
can be achieved simply by using the existing integer variants.

At this point, MySQL enum support is mostly complete,  but there is
still more work needed in the adapter in order to convert data to the
correct types. Once that's done, I'll also be adding some more tests,
but besides that we should be all set.

Change-Id: I670b4879b6ac4753fbfb7052a2f5de83a68b221d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2931
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>