Parse END for Postgres
In Postgres, END, END WORK and END TRANSACTION may be used to commit a
transaction along with COMMIT, COMMIT WORK, and COMMIT TRANSACTION. This
commit updates our parsing logic to check for the incoming dialect, and
account for this difference.

We must process on a per dialect basis in this case because END has a
very different meaning in MySQL. In MySQL, END is used to indicate the
end of a compound statement.

Change-Id: I517f98c1c08ec0d27334d088bc659d834d492966
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2933
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>