replicators: Use "schema" instead of "namespace" throughout
Since mysql accepts the word "SCHEMA" as an alias for "DATABASE", we're
internally using the word "schema" to refer to "the thing that tables
can be qualified by", eg schemas in postgresql and schemas/databases in
mysql. This naming is already used elsewhere, but replicators is an odd
one out for using the word "namespace" instead. This commit fixes that
discrepancy, by updating all the nomenclature in replicators to use
'schema' instead. To avoid a naming conflict, this also changes
`ReplicationAction::SchemaChange` to `ReplicationAction::DdlChange`.

Change-Id: I15124d26c7087e8b351bc552adcd618faab06816
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2936
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>