sql-passes: Table expr aliases take precedence over tables
The proper semantics of unqualified table references in SQL queries are
that references to aliases for table exprs (including CTEs) *within* the
query take precedence over tables in the DDL - this means that a named
CTE should *shadow* a table in the DDL with the same name. This was
implemented incorrectly in the star_expansion pass, where we looked at
tables in the schema first instead of subqueries.

This all comes with a logictest covering this issue.

Fixes: ENG-1702
Change-Id: Ie5e5273aa8d38a96d42fdcc4d3fc69a18bc877d2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2938
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>