Refactor --version logic
This changes how the git commit ID is injected.

Previously, the CARGO_PKG_VERSION environment variable was overwritten
with the commit ID, which was then used by clap as the default version
string. The commit ID came from invoking git via the shell at build
time.

Now, the commit ID is taken from the BUILDKITE_COMMIT environment
variable, which is set automatically by the CI agent.

There is future work, if desired, to provide an automatic solution for
injecting the commit ID for local builds.

Additionally (and unrelated), the explicit name field is removed from
top-level clap attribute. In all cases, it was set to the crate name,
which it defaults to anyway.

Change-Id: I0751158ee80a3f47397a5bbc737c1e564676c899
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2946
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>