readyset-client: Track current schema search path
Add a new `schema_search_path` field to NoriaConnector in
readyset_client, and initialize it at construction time by querying the
relevant upstream from noria_client_adapter.  This is done generically
via a new method on the UpstreamDatabase trait, which for mysql just
looks at whatever database was used to connect, but for postgres
actually makes the query to the upstream database itself (additionally
replacing the special `$user` schema variable with the username used to
connect).

Note that nothing actually *uses* this yet - that'll be coming up next.
Since nothing uses it, it's not easy to automatically test as of this
commit, so I've tested this manually via some debug logs.

Fixes: ENG-1582
Change-Id: Iec85c010dcd3b0bf637b8c81ffdcf59168c64f5a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2947
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>