Add schema_search_path field to ChangeList
As the next step in threading through the current schema search path to
migrations, add a `schema_search_path: Vec<SqlIdentifier>` field to
ChangeList, which is set in migrations based on the currently configured
schema search path in the NoriaConnector in the backend. To keep things
working without making the change huge, this currently just defaults the
field to an empty list, requiring callers to explicitly set it when they
want schemas to be resolved.

To keep encapsulation a little bit cleaner, this refactors ChangeList
itself slightly to define some constructors, getters, and setters for
the bits that should actually be exposed.

As before, nothing actually *reads* this field yet, this change is just
threading it one level deeper into the migration stack.

Fixes: ENG-1709
Change-Id: I6e0f5207abf640b744e3e6928126a632d3ef2925
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2949
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>