clustertests: Allow processes to auto restart
This change extends the clustertest ProcessHandle struct to allow for
automatic restarts. This behavior mimics cluster environments like
kubernetes and allows for future clustertests to test both a cluster
initializing to a health state automatically regarldess of process
startup order as well as automatic recovery from failure scenarios.

A new flag is introduced to the clustertest Deployment construct,
`auto_restart`, which if set true will periodically check the health of
a process and re-run it if it is not running. The default behavior is to
not restart the process because we want this to be an intentional
descision by the testers.

Change-Id: Ied5888faa0dcd5fca47fff04bf24a168ae5b6ef9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2952
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>