docker-compose: Explicitly use pg 14
As of 12c48c4ec (replicators: use pg_logical_emit_message for Postgres
DDL, 2022-07-29), we only support versions of postgresql >= 14 - but
there was no version pinned in the docker-compose file so any machine
that had `postgres:latest` referring to an older version of postgresql
wouldn't work without explicitly running `docker pull postgres`. We are
likely to *change* our support for postgres to support older versions,
but in the meantime let's pin to a specific version in the
docker-compose files to fix the version compatibility issue.

Change-Id: Id2bf9d5754d890ea76aad6afa3f5aa4e2ce7b416
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2953
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>