docs: remove references to display_literals ff
This feature flag was removed in 05714447e (Default to not redacting
sensitive strings, 2022-07-05), but still mentioned in the example
commands in the README.md and development.md. The original behavior that
this flag enabled is now the default behavior, so this commit just
removes the feature flag being set from all the commands in both files.

Change-Id: I2ca26299476304f837fc2f4f93a4348b79afc858
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2954
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI