readyset-client: only report valid SQL proxies
If we fail to parse a query, we proxy it.
SHOW PROXIED QUERIES reports which queries are proxied, including ones
that failed to parse.

Before this change, we were considering any failure to parse in this
list, even if the fallback wasn't able to parse it (it wasn't valid
SQL). This change fixes that so only queries that succeed on the
fallback are reported in this way.

Change-Id: If19adea6a889cf1e0b4a65c18a8980c99a0ef0f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2956
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>