Rewrite queries pre-ExpressionRegistry
Since rewriting will now encode semantic information other than just
that represented in the text of the query, it will be possible that two
queries that're equivalent pre-rewrite end up being non-equivalent
post-rewrite - for example, let's say we have the following tables:

  * `s1.t`
  * `s2.t`

The following query:

    SELECT * FROM t;

might end up being rewritten to either:

    SELECT * FROM s1.t

or:

    SELECT * FROM s2.t

depending on the schema search path (eg if the search path was `["s1",
"s2"]` we'd get the first query, but `["s2", "s1"]` would get the second
query).

Eventually, we'd like to get rid of the ExpressionRegistry entirely (it
kinda just duplicates information that should exist in SqlIncorporator,
and sits at the wrong abstraction level generally) but in the meantime
we need those queries to actually be different from each other. To
achieve that in the fastest way possible, this commit (slightly hackily)
just makes the `SqlIncorporator::rewrite` method public, and calls that
from `Recipe` before attempting to insert into the `ExpressionRegistry`,
instead of inside the various `add_<whatever>` methods in the
`SqlIncorporator`.

Lots of tests (which imo sit at the wrong abstraction level too) had to
change to make this work too - these SqlIncorporator tests have been
having to change a *ton* recently, which I feel like is a huge design
smell -  we should reconsider these tests even really existing,
especially given how many of them there are.

Change-Id: Id07729e81d850fbc9a8968fda81c9ae0d44301b1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2986
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>