Support injecting commit hash into local builds
This change adds a new crate readyset-version, the sole purpose of which
is to get rebuilt whenever the commit ID changes and provide that commit
ID to other crates. It tries to get the commit ID first from the CI
agent via the BUILDKITE_COMMIT environment variable, then from the git
CLI. If both fail, it defaults to "unknown-commit-id".

All usages of BUILDKITE_COMMIT for populating --version information and
logging on application startup are replaced by this new COMMIT_ID
constant from the readyset-version crate.

Fixes: ENG-1712
Change-Id: I26e7225bb8a2549214388b5fca9a8b8cc39262b8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2994
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>