readyset-psql: Support SimpleQueryRow passthrough
We previously prepared and executed simple queries received by the
ReadySet adapter. This CL now makes use of rust-postgres'
`Client::simple_query()` for simple queries received by the adapter and
adds support for passthrough of rust-postgres' SimpleQueryRow.

The passthrough support for SimpleQueryRow is not analagous to our
existing passthrough support for postgres values returned with the
binary protocol. SimpleQueryRow passthrough support is implemented at
the protocol level in readyset-psql by passing the bytes for the row
received from rust-postgres and reconstructing protocol headers as
necessary.

The SimpleQueryRow is written to the client after the entire response is
buffered, which is not necessary. We could improve our performance by
making use of SimpleQueryStream in rust-postgres.

This CL gives us proxy support for multi-statement simple queries. We
could not previously proxy these because they were not supported by the
prepare/execute protocol.

Change-Id: Iccbdc11840821936b036f8e81ab9020939f3fba7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2989
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>