Rewrite COUNT(*) to COUNT(coalesce(col, 0))
Since dataflow requires that all aggregates have an "over" column, we've
always rewritten `COUNT(*)` to `COUNT(col)`, picking a random "bogo"
column from the available columns in source tables as `col`. This works,
except if the `col` we pick has any null values - since `COUNT` only
counts null values, but `COUNT(*)` should count all rows regardless of
nulls, we end up returning the wrong results. To work around this, in
601726832 (Count(*) ignore NULLs, 2021-06-23) we introduced a
non-syntactic `count_nulls` field to FunctionExpr::Count, which was set
to `true` during the rewrite, to tell the count aggregate to count null
values in its over argument. However, doing this non-syntactically means
that if we try to deduplicate queries in the expression registry (which
hashes the string representation of queries to deduplicate) *post*
rewrite, we'll end up thinking `COUNT(*)` and `COUNT(col)` are the same,
even if they're not.

Instead of that, then, this commit makes the semantics change happen in
an entirely syntactic way, by rewriting to `COUNT(coalesce(col, 0))`
instead. This both allows us to remove `count_nulls` everywhere, and
allows the queries to be fully differentiated even post-rewrite.

Change-Id: Id7a0b20895f9d9222cdcd3506b8a4bb0c37a670f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2995
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>